Updated UML Diagram
I included the new Game, Coordinates, and Player classes as well as
cleaned up the diagram and structure. I attempted to add an edit-able
file of the diagram to the repo, but this is a feature that must be paid
for. So I will just continue to update as you guys enter classes and /
or requests