TriviaItem + TriviaFactory
Added the TriviaItem class with children classes TrueFalse and
MultiChoice.  Added TriviaFactory to read from a file and generate
TriviaItem objects.  Also included is a sample TriviaFile.txt containing
sample questions, SampleMain.txt a simple driver i used to test the
functionality of the class, and TriviaItem details.txt which gives an
overview of what the classes are and how they are used.