Trivia.txt formatting
Changed the format required in the Trivia.txt file.
multichoice lines now have the correct answer as the 4th field instead
of the last.  Now a multichoice line looks like
type-category-the question-correct answer-option A-option B-option
C-option D.
true/false lines look the same but instead of a string with either true
or false, there should either be an a for true or b for false.