SQLite Java Database Connector
I added the sqlite jdbc .jar file to the
\GitHub\MacroSoftRepo\Trivia_Maze_Project\CurrentWorkingDirectory\TriviaMaze
directory and set the eclipse project to automatically load it as an
external library.  If it doesnt work for you go to the eclipse project
and click project->properties->libraries tab->Add External JARS
button->navigate to where the .jar is located and select it.