Added a GameGUI class as a proof of concept
Began work on adding GUI component to game
For now, the class just extends the regular Game class so that we can
work simultaneously on the ASCII based game and GUI game.

This commit is more of a proof of concept than anything. Needs a lot of
work. To check it out, uncomment the main method in the GameGUI class,
and when prompted to enter your name, enter anything and then the window
will launch. Not much to see at the moment.