Added temporary code for presentation
For the presentation, I added some temporary code to demonstrate that we
can move around the maze, and win the game. Portions that were added
have been commented as being TEMP CODE.

Also, I added a newline to be printed after the maze, this makes sure
any subsequent printing won't occur on the maze line.

isValidMove has also been changed to reflect the new size of the 3x3
maze. This will need to be redone anyway to reflect the new dynamic size
of the maze.

A printTrophy method has been added to the game class purely for
aesthetic effect. It has a timer delay so it doesn't immediately pop up.