isCorrectAnswer and processAnswer to Maze
Added the isCorrectAnswer and processAnswer methods to the Maze class.

The processAnswer method updates the total questions answered field and
if the answer is correct, increases the questions answered correctly
field as well. Also if the answer is correct unlocks the current door
and moves the player. Only problem I have is I do not know how to get
the direction to move the player in.

If the answer is wrong it increases the total questions answered by the
player and locks the door.