Created method to center GUI's on screen
Created new class MonitorScreen that gives a GUI its origin so that it
will be centered on the screen.

Modified GamePortion to have a default method to center the GUI on the
screen.

Implemented above method for all classes that implement GamePortion

Modified GameDriver to center GUI's on screen after they have been
created.

Note: might be better to change GamePortion to an abstract class and
create a JFrame window class (as everyone has one) and implement the
methods close/centerOnScreen as there are no more GUI classes and
extension is possible.