Modified GameOver ASCII art, and fixed skewness
Created new class, TextAnimation.
---------------------------------------
This class takes care of drawing a string to a JTextArea with a
specified delay between each line. It implements the runnable interface
so it can be used in a new Thread and not interfere with other GUI
component processing.

TextAnimation also provides a method to convert a text file into a
single string.

GameOver
------------------------
*Added constants representing the file names of the winning and losing
ASCII art

*Added new method to streamline process of displaying ASCII art

New Files
---------------------
*Created winningASCIIArt and losingASCIIArt
that holds the winning/losing ASCII art, respectively.