Created QuestionPrompt to handle question/answer
Created QuestionPrompt class that will take care of displaying/answering
questions

This will be a new window that pops up whenever a question needs to be
asked, and allow the Game to have more space for drawing.

Added new method to Maze so it can get the current question's question
type.

Added new constants to TriviaItem to represent the Strings that are the
possible Trivia Question types.

Began redoing GameLogic to use the new QuestionPrompt