Set up new Game Layout and removed older code
Set up the new Game Layout so that the Maze drawing has more space.

Redid how layout was created so that the drawing is in the left of the
layout, and everything else is in the new controlsAndInformation JPanel.

Fixed Maze constants MAX_WIDTH, MAX_HEIGHT to reflect the actualy max
width/height.

Added static methods to Maze class so that you can get the Max Maze
width/height in pixels.