Added Labels for Correct/Wrong Answers and Player Stats
Code added to the Game class to add the labels and handle changing them
when a question is answered. Also changed processAnswer() in the Maze
class to return a boolean so that Game can display the proper label.

Commented my name and a brief message above, I believe, all the code I
added. Read over it and go ahead and remove the comments if it all looks
good!